requirements.txt
streamlit
matplotlib
pandas
futu-api
baostock
akshare
pandas_market_calendars
aiohttp
google-generativeai
dashscope